+++ src/Doc/methods/index.js
@@ -3,11 +3,9 @@ module.exports = Object.assign(
   require('./01-utils'),
   require('./02-accessors'),
   require('./03-match'),
-  require('./04-case'),
-  require('./05-whitespace'),
-  require('./06-tag'),
-  require('./07-loops'),
-  require('./08-lookup'),
+  require('./04-tag'),
+  require('./05-loops'),
+  require('./06-lookup'),
 
   require('./insert/01-replace'),
   require('./insert/02-insert'),
@@ -20,6 +18,8 @@ module.exports = Object.assign(
   require('./transform/01-sort'),
   require('./transform/02-normalize'),
   require('./transform/03-split'),
-  require('./transform/04-join'),
-  require('./transform/05-contract')
+  require('./transform/04-case'),
+  require('./transform/05-whitespace'),
+  require('./transform/06-join'),
+  require('./transform/07-contract')
 )