+++ tests/punctuation.test.js
@@ -17,21 +17,29 @@ test('normalize question mark', function(t) {
   t.end()
 })
 
-// test('normalize unicode', function(t) {
-//   const str = `• Spencęr & JOhn™ ⟨lmt⟩.`;
-//   const doc = nlp(str);
-//   t.equal(doc.text(), str, 'text out');
-//   t.equal(doc.text(), 'spencer & john lmt', 'normal out');
-//   t.end();
-// });
+test('hyphenated', function(t) {
+  let doc = nlp('and check this out! a walk-in microwave.')
+  doc.hyphenated().deHyphenate()
+  t.equal(doc.text(), 'and check this out! a walk in microwave.', 'dehyphenate')
+  t.end()
+})
 
-// test('normalize quotes ', function(t) {
-//   const str = `،one’ «two» ‘three’ “four” 'five' "six."`
-//   const doc = nlp(str)
-//   t.equal(doc.text(), str, 'text out-3')
-//   t.equal(doc.text(), 'one two three four five six.', 'normal out-3')
-//   t.end()
-// })
+test('normalize unicode', function(t) {
+  // const str = `• Spencęr & JOhn™ ⟨lmt⟩.`
+  const str = ` Spencęr & JOhn™ ⟨lmt⟩.`
+  const doc = nlp(str)
+  t.equal(doc.text(), str, 'text out')
+  t.equal(doc.text('reduced'), 'spencer & john lmt', 'normal out')
+  t.end()
+})
+
+test('normalize quotes ', function(t) {
+  const str = `،one’ «two» ‘three’ “four” 'five' "six."`
+  const doc = nlp(str)
+  t.equal(doc.text(), str, 'text out-3')
+  t.equal(doc.text('clean'), 'one two three four five six.', 'normal out-3')
+  t.end()
+})
 
 test('toParentheses', function(t) {
   let doc = nlp(`you could still go to McGill, the Harvard of Canada!`)